Move HttpClientMock to new testutils file
