+++ test/test_pairing.py
@@ -1,29 +1,8 @@
 import toopher
 import unittest
-import requests
 import uuid
 import json
-
-class HttpClientMock(object):
-    def __init__(self, paths):
-        self.paths = paths
-
-    def request(self, method, uri, data=None, headers=None, params=None):
-        self.last_called_uri = uri
-        self.last_called_method = method
-        self.last_called_data = data if data else params
-        self.last_called_headers = headers
-
-        uri = uri.split(toopher.DEFAULT_BASE_URL)[1][1:]
-        return ResponseMock(self.paths[uri])
-
-
-class ResponseMock(requests.Response):
-    def __init__(self, response):
-        self.encoding = 'utf-8'
-        self.status_code = int(response[0])
-        self._content = response[1]
-
+from .testutils import HttpClientMock
 
 class PairingTests(unittest.TestCase):
     toopher.DEFAULT_BASE_URL = 'https://api.toopher.test/v1'
